Refactor corrections
